refactor(e2e): real ada test target
Define a new testing target that executes only
tests suitable for networks with real ADA (mainnet)
