+++ packages/e2e/jest.config.js
@@ -6,6 +6,19 @@ const project = (displayName) => ({
   transform: { '^.+\\.test.ts?$': 'ts-jest' }
 });
 
+// jest tests suitable to run in an environment with real ADA (not tADA)
+const realAdaTestFileNames = [
+  'delegation',
+  'metadata',
+  'mint',
+  'multisignature',
+  'nft',
+  'pouchDbWalletStores',
+  'txChainHistory',
+  'txChaining',
+  'unspendableUtxos'
+];
+
 module.exports = {
   projects: [
     { ...project('blockfrost'), globalSetup: './test/blockfrost/setup.ts' },
@@ -16,7 +29,14 @@ module.exports = {
     project('ogmios'),
     project('projection'),
     project('providers'),
-    project('wallet')
+    project('wallet'),
+    {
+      displayName: 'wallet-real-ada',
+      preset: 'ts-jest',
+      setupFiles: ['dotenv/config'],
+      testMatch: [`<rootDir>/test/wallet/SingleAddressWallet/(${realAdaTestFileNames.join('|')}).test.ts`],
+      transform: { '^.+\\.test.ts?$': 'ts-jest' }
+    }
   ],
   testTimeout: 1000 * 60 * 25
 };
